+++ frontend/src/components/layout/Main.tsx
@@ -1,6 +1,6 @@
-import { AnimatePresence, motion } from "framer-motion";
 import type React from "react";
 import { memo, useMemo } from "react";
+import { FadeIn } from "../ui";
 import { NavBar } from "./NavBar";
 
 export type MainProps = {
@@ -30,17 +30,11 @@ export const Main: React.FC<MainProps> = memo(
         style={style}
       >
         {showNavBar && <NavBar />}
-        <AnimatePresence>
-          <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            transition={{ duration: 1 }}
-            className="hero-content w-full"
-          >
-            <div className="max-w-6xl w-full">{children}</div>
-          </motion.div>
-        </AnimatePresence>
+        <div className="hero-content w-full">
+          <div className="max-w-6xl w-full">
+            <FadeIn>{children}</FadeIn>
+          </div>
+        </div>
       </main>
     );
   },